+++ R/typst.R
@@ -105,7 +105,11 @@ typst_cell_options <- function(ht, i, j, rs, cs, al, row_h) {
   opts <- c()
   if (rs > 1) opts <- c(opts, sprintf("rowspan: %d", rs))
   if (cs > 1) opts <- c(opts, sprintf("colspan: %d", cs))
-  if (!is.na(al)) opts <- c(opts, sprintf("align: %s", al))
+
+  va <- c(top = "top", middle = "center", bottom = "bottom")[valign(ht)[i, j]]
+  if (!is.na(al) || !is.na(va)) {
+    opts <- c(opts, sprintf("align: (%s, %s)", al, va))
+  }
 
   if (!is.na(row_h)) {
     rh <- row_h
@@ -161,8 +165,14 @@ typst_cell_text <- function(ht, i, j, cell_text) {
   if (!is.na(f <- font(ht)[i, j])) text_opts <- c(text_opts, sprintf("family: \"%s\"", f))
 
   if (length(text_opts) > 0) {
-    sprintf("#text(%s)[%s]", paste(text_opts, collapse = ", "), cell_text)
+    text <- sprintf("#text(%s)[%s]", paste(text_opts, collapse = ", "), cell_text)
   } else {
-    cell_text
+    text <- cell_text
   }
+
+  if (!wrap(ht)[i, j]) {
+    text <- sprintf("#box(breakable: false)[%s]", text)
+  }
+
+  text
 }